feat: Split and improve code
